backend/backend/requirements.txt
fastapi
pydantic
spacy
translate
uvicorn
google-generativeai
